+++ nbproject/project.properties
@@ -80,7 +80,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=ProyectoCS
-main.class=proyectocs.Login
+main.class=VISTAS.Login
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
